requirements.txt
numpy==1.25.2
pandas==2.1.1
requests==2.31.0
duckdb==1.1.3
python-dotenv==1.0.0